Support target selector for more navi apps (fix #13540)
