+++ main/src/main/java/cgeo/geocaching/apps/navi/OruxMapsApp.java
@@ -43,7 +43,7 @@ abstract class OruxMapsApp extends AbstractPointNavigationApp {
     }
 
     @Override
-    public void navigate(@NonNull final Context context, @NonNull final Geocache cache) {
+    public void navigateWithoutSelector(@NonNull final Context context, @NonNull final Geocache cache) {
         final Geopoint coords = cache.getCoords();
         assert coords != null; // guaranteed by caller
         navigate(context, coords, cache.getName());
